PSUSensor: Fix error for decimal part of scalefactor
Customized scale factor declaration from schema in entity-manager,
declared as type "number." If a number is used in PSUSensor, it loses
the decimal part when transforming the value from variant to unsigned
int, resulting in incorrect calculations.

For example, a sensor named "ADC_TEST" is set as below:

{
    "Address": 31,
    "Bus": 31,
    "Labels": [
        "in6"
    ],
    "Name": "FANBOARD_ADC",
    "Type": "ADC128D818",
    "in6_Name": "ADC_TEST",
    "in6_Scale": 31.9458
}

but scale factor in debug print is 31, not 31.9458:

Jul 28 22:50:36 bmc psusensor[788]: DEBUG_FACTOR:31

The more reasonable way to resolve the issue is to use a variant to
float or double, as adcsensor has been doing so far.

Tested:
Get adc128d818 and ina233 reading from PSUSensor by dbus introspect

+++ src/PSUSensor.hpp
@@ -23,7 +23,7 @@ class PSUSensor : public Sensor, public std::enable_shared_from_this<PSUSensor>
               std::vector<thresholds::Threshold>&& thresholds,
               const std::string& sensorConfiguration,
               const PowerState& powerState, const std::string& sensorUnits,
-              unsigned int factor, double max, double min, double offset,
+              double factor, double max, double min, double offset,
               const std::string& label, size_t tSize, double pollRate,
               const std::shared_ptr<I2CDevice>& i2cDevice);
     ~PSUSensor() override;
@@ -48,7 +48,7 @@ class PSUSensor : public Sensor, public std::enable_shared_from_this<PSUSensor>
     boost::asio::random_access_file inputDev;
     boost::asio::steady_timer waitTimer;
     std::string path;
-    unsigned int sensorFactor;
+    double sensorFactor;
     double sensorOffset;
     thresholds::ThresholdTimer thresholdTimer;
     void restartRead();
@@ -67,7 +67,7 @@ class PSUSensor : public Sensor, public std::enable_shared_from_this<PSUSensor>
 class PSUProperty
 {
   public:
-    PSUProperty(std::string name, double max, double min, unsigned int factor,
+    PSUProperty(std::string name, double max, double min, double factor,
                 double offset) :
         labelTypeName(std::move(name)),
         maxReading(max), minReading(min), sensorScaleFactor(factor),
@@ -78,6 +78,6 @@ class PSUProperty
     std::string labelTypeName;
     double maxReading;
     double minReading;
-    unsigned int sensorScaleFactor;
+    double sensorScaleFactor;
     double sensorOffset;
 };
